+++ infra/resources/_modules/key_vault/locals.tf
@@ -5,7 +5,6 @@ locals {
       azure_client_secret_credential_secret    = "AZURE-CLIENT-SECRET-CREDENTIAL-SECRET"
       azure_client_secret_credential_client_id = "AZURE-CLIENT-SECRET-CREDENTIAL-CLIENT-ID"
       bo_auth_session_secret                   = "bo-auth-session-secret"
-      function_apim_key                        = "${var.prefix}-${var.env_short}-services-cms-webapp-fn-apim-key"
       jira_token                               = "JIRA-TOKEN"
       legacy_cosmosdb_connectionstring         = "legacy-cosmosdb-connectionstring"
       legacy_cosmosdb_key                      = "legacy-cosmosdb-key"