+++ packages/external-clients/src/apim-client/index.ts
@@ -471,6 +471,9 @@ const upsertSubscription = (
             scope: `/products/${productId}`,
             state: "active",
           },
+          {
+            ifMatch: "*",
+          },
         ),
       identity,
     ),